// src/interfaces/params.interface.ts
export interface PageParams {
	pageNum?: number
	pageSize?: number
	keyword?: string
}

export interface fineuploadParams {
	partIndex: number
	partSize: number
	totalParts: number
	totalFileSize: number
	filename: number
	uuid: number
	file: string
}
export interface sceneListParams {
	sceneId?: string
	sceneName?: string
	srcType?: string
}

export interface brokenLineParams {
	obj: string
	queryType: number
	startTimeStr: string
	endTimeStr: string
	topicId?: string
	tempType?: string
	timeType?: string
	timeStr?: string
	companyId?: string
}
// ...
export interface tempListParams {
	userId?: string
	keyword: string
	companyId?: string
	topicId?: string
}
// ...
export interface businessListParams {
	name: string
}

export interface addBusiness {
	id?: string
	name: string
	username: string
	password: string
	domainName: string
	contact: string
	phone: string
	emailBox: string
	companyPhone: string
	companyAddress: string
	address: string
	companyStatus: string
}

export interface createResourceParams {
	companyId: string
	resourceId?: string
	vrCount: number
	expireDate: number
	areaBindList: areaListItem[]
}
export interface addTheme {
	id?: string
	companyId: string
	name: string
	simpleName: string
	domainName: string
	logo: string
}
export interface topicTempListParams {
	keyword?: string
	styleId?: string
	typeId?: string
}
export interface tempAdd {
	topicId: string
	topicTempList: object[]
}
export interface tempEdit {
	tempId: string
	name: string
	createType: number
	publicAttr: boolean
	userList: []
}

export interface topicUserParams {
	keyword?: string
	areaId?: string
	startTime?: string
	endTime?: string
	topicId: string
}
export interface topicUserAdd {
	id?: string
	companyId?: string
	topicId?: string
	username?: string
	telephone?: string
	releases?: number
	areaList: object[]
}

export interface businessHallParams {
	userId: string
	companyId: string
	topicId: string
	tempId: string
	tempName: string
}

export interface cloneHallParams {
	userId: string
	companyId: string
	topicId: string
	tempId: string
	tempName: string
}