// src/interfaces/api.interface.ts
export interface baseRes<D = {}> {
	success: boolean
	code: number
	message: string
	data: D //返回的数据字段
}

export interface PageData<D = {}> {
	tempGroup: any
	bindList: any
	filePreviewUrl: string
	fileSaveUrl: string
	currentPage: number
	pageNum: number
	pageSize: number
	totalPage: number
	count: number
	list: D[]
	fileHost: string
	helpVideo: {
		videoUrl: string
		coverUrl: string
	}
}

export interface callDataItem {
	todayCount: number
	allCount: number
}

export interface seriesListItem {
	seriesData: number[]
	seriesDataBig?: number[]
	seriesName: string
}

export interface brokenLineItem {
	xaxisData: string[]
	seriesList: seriesListItem[]
}

export interface countProvinceItem {
	countId: number
	province: string
	ratio: number
	align?: string
}

export interface countCountryItem {
	countId: number
	country: string
	ratio: number
}

export interface countDeviceItem {
	countId: number
	deviceOS: string
	ratio: number
}

export interface tempListItem {
	id: string
	name: string
	tempName: string
	sz: string
	validSz: string
}
// ...
export interface platformsItem {
	key: number | string
	value: string
}
// ...
export interface userListItem {
	birthday: string
	email: string
	id: string
	idCard: string
	nickname: string
	realName: string
	roleList: null
	sex: number
	state: number
	telephone: string
	userOrigin: null
	username: string
	name?: string
}
export interface styleList {
	styleId?: string
}
export interface typeList {
	typeId?: string
}
export interface hallInfo {
	belonger?: string
	check: number
	checkNote: null | string
	endTime?: number
	durationEndTs?: number
	extData?: string
	lockY: null | boolean
	lockYAngle: string
	lockZoom: number
	styleId?: string
	tempId: string
	tempName: string
	typeId?: string
	userId?: string
	userName?: string
	varSetting?: number
	addViewCount?: number
	addLikeCount?: number
	modelCount?: number
	gifCount?: number
	styleList?: styleList[]
	typeList?: typeList[]
	viewCount?: number
	likeCount?: number
	fastLayout: boolean
	sceneId: string
	sceneName?: string
	floorSetting: boolean
}

export interface hallTableInfo {
	tempName: string
	validSz: number
	styleList: styleListItem2[]
	typeList: typeListItem2[]
	createTs: number
	durationEndTs: number
	userName: string
	userPhone: string
	sceneFromVal: string
	boutique: boolean
	aliStatic: number
	fastLayout: boolean
	renderFlowStatus: number
	renderOver: null | boolean
	extData?: string
	parentId?: string | null
	parentName?: string | null
	sceneId: string
	sceneName: string
	tempId: string
	publishAfterStatus: number
	areaRemark?: string
}
// ...
export interface userDataItem {
	allUserCount: number
	registerRate: number
	userCount: number
	userExtCount: number
	userExtRate: number
	userPublishCount: number
	userPublishRate: number
	userTempCount: number
	userTempRate: number
	visitorCount: number
}

export interface userDataItem2 {
	aliStatic: number
	boutiqueCount: number
	notPublishCount: number
	publishCount: number
	tempCount: number
}
// ...
export interface styleListItem2 {
	styleId: string
	styleName: string
	tempId: string
}

export interface typeListItem2 {
	tempId: string
	typeId: string
	typeName: string
}
// ...
export interface businessInfo {
	name: string
	password: string
	repassword?: string
	username: string
	phone: number
	companyPhone: number
	companyAddress: string
	companyStatus: boolean
	emailBox: string
	avatar: string
	logo: string
}
export interface resourceInfoItem {
	useTempCount: number
	tempCount: number
	useVrCount: number
	companyId: string
	id: string
	vrCount: number
	expireDate: number
	areaList: areaListItem[]
}

export interface areaListItem {
	areaId: number
	remark: string
	min: number
	max: number
	price: number
	tempCount: number
	useTempCount: number
	addCount?: number
}

export interface resourceListItem {
	id: string
	orderNumber: string
	resourceType: number
	tempCount: number
	vrCount: number
	createTs: number
	expireDate: number
}

export interface getResourceInfoItem {
	companyId: string
	id: string
	useVrCount: number
	vrCount: number
	addVrCount?: number
	createTs: number
	expireDate: number
	areaList: areaListItem[]
	tempCount: number
	resourceCount: number
	resourceType: number
	addCountSum?: number
}
export interface areaInterva {
	entityPrice?: string
	areaId: number
	max?: number
	min?: number
	price?: number
	remark: string
	virtualPrice?: string
	userRemainCount?: number
	remainCount?: number
}

export interface tempListItem {
	id: string
	companyId: string
	tempId: string
	name: string
	createType: number
	createTs: string
	createdBy: string
	createdName: string
	roamCount: number
	thumb: string
	validSz: string
	userId: string
	userName: string
}

export interface tempKeywordItem {
	id: string
	name: string
	tempName: string
	sz: string
	validSz: string
}
export interface bestTemp {
	id: string
	name: string
	sz: string
	validSz: string
	thumb: string
}

export interface listData<D = {}> {
	currentPage: number
	pageSize: number
	totalPage: number
	count: number
	list: D[]
}
export interface topicTempInfo {
	id: string
	name: string
	publicAttr: boolean
	tempId: string
	userList: topicUserItem[]
}
export interface topicUserItem {
	id: string
	username: string
	nickname: string
	publishCount?: number
	sceneTotalCount?: number
	releases?: number
	areaList?: areaInterva[]
}
export interface themeInfo {
	name: string
	logo: string
	id: string
}

export interface businessListItem {
	id: string
	name: string
}

export interface tempPageListItem {
	pageNum: number
	pageSize: number
	companyId: string
	topicId: string
	tempName: string
	tempId: string
	phoneOrName: string
	aliStatic: number
	isCheck: number
	startTime: string
	endTime: string
	valueAddedService: number
}

export interface topicTempItem {
	tempId: string
	tempName: string
	topicId: string
}

export interface topicUserListItem {
	id: string
	nickname: string
	username: string
	topicId: string
}

export interface serviceConfigItem {
	tempId: string
	varLook: number
	varLookCount: number
	varLookInfo: varLookInfoItem
	myBrowseService: number
}

export interface varLookInfoItem {
	expireFlag: boolean
	remainCount: number
	totalCount: number
	useCount: number
}

export interface polymerizationInfo {
	imgUrl: string
}

export interface loginListItem {
	id: string
	name: string
	thumb: string
	styleFile: string
	relatedTopic: boolean
	rackStatus: boolean
	sort: number
}